Refactor oauth components & merge with Webhook component
