+++ frontend/package.json
@@ -47,16 +47,10 @@
     "gulp-strip-debug": "^1.1.0",
     "gulp-uglify": "^2.0.0",
     "gulp-util": "^3.0.7",
-    "immutability-helper": "^2.1.1",
-    "js-cookie": "^2.1.3",
-    "moment": "^2.17.1",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.2",
     "react-modal": "^1.6.5",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
-    "vinyl-transform": "^1.0.0",
-    "whatwg-fetch": "^1.0.0"
+    "vinyl-transform": "^1.0.0"
   },
   "browserify": {
     "transform": [
