+++ frontend/src/react/components/app/defaultHeaders.js
@@ -1,0 +1,6 @@
+import Cookies from 'js-cookie';
+
+export default {
+  'Content-Type': 'application/x-www-form-urlencoded',
+  'X-CSRFToken': Cookies.get('csrftoken')
+};
