+++ frontend/src/react/components/appList.jsx
@@ -4,379 +4,6 @@ import 'whatwg-fetch';
 import Modal from 'react-modal';
 import {App} from './app.jsx';
 import {AppForm} from './appForm.jsx';
-import Collapse, { Panel } from 'rc-collapse';
-
-import Webhook from './webhook.jsx';
-
-class App extends React.Component {
-  constructor(props){
-    super(props);
-    this.state = {
-      editing: false,
-      copied: false,
-      error: ''
-    };
-    this.changeName = this.changeName.bind(this);
-    this.editName = this.editName.bind(this);
-    this.regenToken = this.regenToken.bind(this);
-    this.regenConfirm = this.regenConfirm.bind(this);
-    this.deleteApp = this.deleteApp.bind(this);
-    this.deleteConfirm = this.deleteConfirm.bind(this);
-    this.stopEditing = this.stopEditing.bind(this);
-    this.copyToken = this.copyToken.bind(this);
-    this.setCopyText = this.setCopyText.bind(this);
-
-    /*
-      Make moment.js say 'just now' if the time difference is less
-      than five seconds either way to get round non-sync'd server and
-      client time.
-      This may make 'a few seconds ago' somewhat redundant, but it's
-      worth it.
-    */
-    moment.fn.fromNowOrNow = function(a) {
-      if (Math.abs(moment().diff(this)) < 5000) {
-        return 'just now';
-      }
-      return this.fromNow(a);
-    };
-
-  }
-
-  changeName(e){
-    e.preventDefault();
-    let that = this;
-    fetch('/dashboard/api/rename/', {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'X-CSRFToken':Cookies.get('csrftoken')
-      },
-      body: 'new_name=' + this.refs.name.value +
-        '&app_id=' + this.props.appId
-    }).then((res)=>{
-      if(res.ok){
-        return res.json();
-      } else {
-        throw new Error('An error occured');
-      }
-    }).then((json)=>{
-      if(json.success){
-        let newApp = {
-          name: that.refs.name.value,
-          id: that.props.appId,
-          token: that.props.appKey,
-          created: that.props.created,
-          updated: json.date
-        };
-        that.props.update(that.props.appId, newApp);
-        that.refs.name.value = '';
-        that.setState({
-          editing: false,
-          error: ''
-        });
-      }else{
-        throw new Error(json.message);
-      }
-    }).catch((err)=>{
-      that.setState({
-        error: err.message
-      });
-    });
-  }
-
-  regenToken(){
-    let that = this;
-    fetch('/dashboard/api/regen/', {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'X-CSRFToken':Cookies.get('csrftoken')
-      },
-      body: 'app_id=' + this.props.appId
-    }).then((res)=>{
-      if(res.ok){
-        return res.json();
-      } else {
-        throw new Error('An error occured');
-      }
-    }).then((json)=>{
-      if(json.success){
-        let newApp = {
-          name: that.props.name,
-          id: that.props.appId,
-          token: json.app.token,
-          created: that.props.created,
-          updated: json.app.date
-        };
-        that.props.update(that.props.appId, newApp);
-        that.setState({
-          error:''
-        });
-      }else{
-        throw new Error(json.message);
-      }
-    }).catch((err)=>{
-      that.setState({
-        error: err.message
-      });
-    });
-  }
-
-  deleteApp(){
-    let that = this;
-    fetch('/dashboard/api/delete/', {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'X-CSRFToken': Cookies.get('csrftoken')
-      },
-      body: 'app_id=' + this.props.appId
-    }).then((res)=>{
-      if(res.ok){
-        return res.json();
-      }else{
-        throw new Error('An error occured');
-      }
-    }).then((json)=>{
-      if(json.success){
-        that.props.remove(that.props.appId);
-      }else{
-        throw new Error(json.message);
-      }
-    }).catch((err)=>{
-      that.setState({
-        error: err.message
-      });
-    });
-  }
-
-  deleteConfirm(e){
-    e.preventDefault();
-    if(confirm('Are you sure you want to delete this app?')){
-      this.deleteApp();
-    }
-  }
-  regenConfirm(e){
-    e.preventDefault();
-    if(confirm('Are you sure you want to regenerate your api token?')){
-      this.regenToken();
-    }
-  }
-
-  editName(e){
-    e.preventDefault();
-    this.setState({
-      editing: true
-    });
-  }
-
-  stopEditing(e){
-    e.preventDefault();
-    this.setState({
-      editing: false
-    });
-  }
-
-  copyToken(e){
-    e.preventDefault();
-
-    let tokenElement = this.refs.apiToken;
-    tokenElement.select();
-
-    try {
-      // copy text
-      document.execCommand('copy');
-      this.setState({
-        copied: true
-      });
-      tokenElement.blur();
-    }catch (err) {
-      alert('please press Ctrl/Cmd+C to copy');
-    }
-  }
-
-  setCopyText(){
-    this.setState({
-      copied: false
-    });
-  }
-
-
-  render () {
-    const { webhook } = this.props;
-    return <div className="app pure-u-1 pure-u-xl-1-2">
-      <div className="card">
-        {this.state.editing ? (
-          <form className="pure-form" onSubmit={this.changeName}>
-            <fieldset>
-              <input type="text" autoFocus placeholder={this.props.name} ref="name"/>
-              <button type="submit" className="pure-button pure-button-primary padded" onClick={this.changeName}>Submit</button>
-              <button className="pure-button button-error padded" onClick={this.stopEditing}>Cancel</button>
-            </fieldset>
-          </form>
-        ):(
-          <div className="pure-g">
-            <div className="pure-u-1-2">
-              <h2>{this.props.name}</h2>
-            </div>
-            <div className="pure-u-1-2 flexTopRight">
-              <button className="pure-button pure-button-primary padded" onClick={this.editName}>
-                <i className="fa fa-pencil" aria-hidden="true"></i>
-              </button>
-              <button className="pure-button button-error padded" onClick={this.deleteConfirm}>
-                <i className="fa fa-trash-o" aria-hidden="true"></i>
-              </button>
-            </div>
-          </div>
-        )}
-        <div className="pure-g">
-          <div className="pure-u-1">
-            API Token
-            <form className="pure-form pure-g">
-              <div className="pure-u-2-3">
-                <input
-                  type="text"
-                  ref="apiToken"
-                  className="pure-input-1"
-                  value={this.props.appKey}
-                  readOnly
-                  style={{ 'borderRadius': '4px 0px 0px 4px'}}
-                />
-              </div>
-              <div className="pure-u-1-6">
-                <button
-                  className="pure-button pure-button-primary pure-input-1 tooltip"
-                  onClick={this.copyToken}
-                  onMouseEnter={this.setCopyText}
-                  style={{ 'border': '1px solid #ccc', 'borderRadius': '0px'}}
-                >
-                  <i className="fa fa-clipboard" aria-hidden="true"></i>
-                  <span>{this.state.copied?'Copied!':'Click to copy to clipboard'}</span>
-                </button>
-              </div>
-              <div className="pure-u-1-6">
-                <button
-                  className="pure-button pure-button-primary pure-input-1"
-                  onClick={this.regenConfirm}
-                  style={{ 'border': '1px solid #ccc', 'borderRadius': '0px 4px 4px 0px'}}
-                >
-                  <i className="fa fa-refresh" aria-hidden="true"></i>
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-        <p title={moment(this.props.created).format('dddd, Do MMMM YYYY, h:mm:ss a')}>Created: {moment(this.props.created).fromNowOrNow()}</p>
-        <p title={moment(this.props.updated).format('dddd, Do MMMM YYYY, h:mm:ss a')}>Updated: {moment(this.props.updated).fromNowOrNow()}</p>
-        <Collapse>
-          <Panel
-            header="Webhook Settings"
-            showArrow
-          >
-            <Webhook
-              appId={this.props.appId}
-              {...webhook}
-            />
-          </Panel>
-        </Collapse>
-
-        <label className="error">{this.state.error}</label>
-        </div>
-    </div>;
-  }
-}
-
-App.propTypes = {
-  name: React.PropTypes.string.isRequired,
-  appId: React.PropTypes.string.isRequired,
-  appKey: React.PropTypes.string.isRequired,
-  created: React.PropTypes.string.isRequired,
-  updated: React.PropTypes.string.isRequired,
-  webhook: React.PropTypes.shape({
-    verification_secret: React.PropTypes.string.isRequired,
-    url: React.PropTypes.string.isRequired,
-    siteid: React.PropTypes.string.isRequired,
-    roomid: React.PropTypes.string.isRequired,
-    contact: React.PropTypes.string.isRequired,
-  }),
-  update: React.PropTypes.func.isRequired,
-  remove: React.PropTypes.func.isRequired
-};
-
-class AppForm extends React.Component {
-  constructor(props){
-    super(props);
-    this.state = {
-      error: ''
-    };
-    this.submitForm = this.submitForm.bind(this);
-  }
-
-  submitForm(e){
-    e.preventDefault();
-    let that = this;
-    fetch('/dashboard/api/create/', {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'X-CSRFToken':Cookies.get('csrftoken')
-      },
-      body: 'name=' + this.refs.name.value
-    }).then((res)=>{
-      if(res.ok){
-        return res.json();
-      } else {
-        throw new Error('An error occured');
-      }
-    }).then((json)=>{
-      if(json.success){
-        let newApp = json.app;
-        newApp['name'] = that.refs.name.value;
-        that.refs.name.value = '';
-        that.props.add(newApp);
-        that.props.close();
-      }else{
-        throw new Error(json.message);
-      }
-    }).catch((err)=>{
-      that.setState({
-        error: err.message
-      });
-    });
-  }
-  render () {
-    return <div className="appForm">
-      <h2>Create App</h2>
-      <form className="pure-form pure-form-stacked" onSubmit={this.submitForm}>
-        <fieldset>
-          <div className="pure-g">
-            <div className="pure-u-1">
-              <label htmlFor="name">App Name</label>
-              <input autoFocus id="name" ref="name" className="pure-u-1" type="text"/>
-            </div>
-          </div>
-          <div className="pure-g">
-            <div className="pure-u-1-24"></div>
-            <button type="submit" className="pure-button pure-button-primary pure-u-10-24">Create</button>
-            <div className="pure-u-2-24"></div>
-            <button type="button" className="pure-button button-error pure-u-10-24" onClick={this.props.close}>Cancel</button>
-            <div className="pure-u-1-24>"></div>
-          </div>
-          <label className="error">{this.state.error}</label>
-        </fieldset>
-      </form>
-    </div>;
-  }
-}
-
-AppForm.propTypes = {
-  add: React.PropTypes.func.isRequired,
-  close: React.PropTypes.func.isRequired
-};
 
 class AppList extends React.Component {
   constructor (props) {
@@ -452,10 +79,7 @@ class AppList extends React.Component {
             key={i}
             update={this.updateApp}
             remove={this.deleteApp}
-            clientId={app.oauth.client_id}
-            clientSecret={app.oauth.client_secret}
-            callbackUrl={app.oauth.callback_url}
-            scopes={app.oauth.scopes}
+            oauth={app.oauth}
             webhook={app.webhook}
           />;
         })}
@@ -464,8 +88,8 @@ class AppList extends React.Component {
         isOpen={this.state.showCreate}
         contentLabel="Create app form"
         onRequestClose={this.hideForm}
-				className="Modal"
-				overlayClassName="Overlay"
+        className="Modal"
+        overlayClassName="Overlay"
       >
         <AppForm add={this.addApp} close={this.hideForm}/>
       </Modal>
